influencer_roi/requirements.txt
# Django Backend
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8

# Streamlit Frontend
streamlit==1.28.1
plotly==5.17.0
pandas==2.2.0
numpy==1.26.2
requests==2.31.0
streamlit-aggrid==0.3.4

# Development
python-dotenv==1.0.0 